Generate & attach PDF cert. to confirmation email
https://eaflood.atlassian.net/browse/WC-479

We attach PDF certificates to new registration emails. However, for renewals, we currently ask them to log in.

If the user can't log in and renewed through an AD process instead, they have no way to access the certificate.

This change will generate and attach the updated certificate to the "renewal complete" email so no user needs to log in to access it.

🔥 WIP 🔥