Initial working version of PDF generation
Simply a couple of tests to ensure we can generate a PDF without error. Main problem to overcome was the fact the bash script we set as the exe for WickedPdf is sat in an engine, whereas previous services this would be sat in the app.

Using `WasteCarriersEngine::Engine.root` instead of `Rails.root` seems to solve it, but will only be sure when we call it from a host app.