Let users proceed to renewal_received after pending WP payment
https://eaflood.atlassian.net/browse/WC-409

Sometimes WorldPay can't approve a payment immediately, although it may be approved after a delay.

In this case, we should let the user continue with the process as they will likely be charged eventually. However, as we don't want the renewal to be completed until payment has been taken, we should send it to the 'renewal received' state until a back office user records the payment.